Added robots.txt functionality
In a multi-domain environment, you often need different robots.txt files. Making them editable solves both this problem and the SEO editor's suffering of never being able to change this file.

Tested different scenarios, but this one with CodeEditor fits best - it allows the editor to work on larger robots.txt files in a text-editor fashion.